pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "abatement-crawler"
version = "0.1.0"
description = "Webcrawler to find data on carbon abatement measures"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "anthropic>=0.25",
    "httpx>=0.27",
    "trafilatura>=1.8",
    "beautifulsoup4>=4.12",
    "pymupdf>=1.24",
    "pdfplumber>=0.11",
    "pandas>=2.0",
    "pyarrow>=15.0",
    "python-docx>=1.1",
    "pint>=0.23",
    "pyyaml>=6.0",
    "python-slugify>=8.0",
    "requests>=2.31",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[project.scripts]
abatement-crawler = "abatement_crawler.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["abatement_crawler*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
